+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Version 0.19.0
+# Version 0.18.6
 * Fixed PlanTotem not recognizing fully supported plans as supported
 * Fixed ItemStand and ArmorStand losing attached items after a reload when placed via direct build
 * Fixed ItemStand and ArmorStand items not being captured correctly when saving a new blueprint